+++ tests/test_pipe.py
@@ -1,4 +1,10 @@
-from pypsexec.pipe import FSCTLPipeWait
+# -*- coding: utf-8 -*-
+# Copyright: (c) 2019, Jordan Borean (@jborean93) <jborean93@gmail.com>
+# MIT License (see LICENSE or https://opensource.org/licenses/MIT)
+
+from pypsexec.pipe import (
+    FSCTLPipeWait,
+)
 
 
 class TestFSCTLPipeWait(object):