+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0 - TBD
 
 * Dropped support for Python 2.6
+* Updated the `PAExec` executable to `1.27`
 * Set minumum version of smbprotocol to `1.0.0` to take advantage of new simpler library.